usage_monitoring/requirements.txt
pyrebase4==4.5.0
flask-cors==3.0.10
pyparsing<3,>=2.0.2
